Split ev/core/memory.py into a package by domain
Same extract-and-recompose approach as the tools.py split: Memory's
30+ CRUD domains become mixins composed in memory/base.py, with
memory/__init__.py re-exporting Memory so every existing import path
and call site keeps working unchanged. No behavior change, no schema
changes.
